refactor(api): split login from user service
