+++ packages/api/src/authentication/express.auth.hooks.test.ts
@@ -1,11 +1,11 @@
-/* eslint-disable @typescript-eslint/no-explicit-any */
 import passport = require("passport");
 import { Express } from "express";
 import passportLocal = require("passport-local");
 import passportJwt = require("passport-jwt");
-import userService from "../../src/modules/user/user.service";
-import { authMocks } from "../../src/authentication/express.auth.hooks";
+import userService from "../modules/user/user.service";
+import { authMocks } from "./express.auth.hooks";
 import { ObjectId } from "mongodb";
+import userAuthService from "../modules/user/services/auth/user.auth.service";
 
 describe("express.auth.hooks", () => {
     let passportMock: jest.SpyInstance;
@@ -35,7 +35,7 @@ describe("express.auth.hooks", () => {
 
             jest.spyOn(passportLocal, "Strategy").mockImplementation(strat as any);
             // @ts-expect-error: mock user
-            jest.spyOn(userService, "login").mockImplementation(email =>
+            jest.spyOn(userAuthService, "login").mockImplementation(email =>
                 Promise.resolve({
                     _id: new ObjectId(),
                     email,
@@ -69,8 +69,7 @@ describe("express.auth.hooks", () => {
             }
 
             jest.spyOn(passportLocal, "Strategy").mockImplementation(strat as any);
-            // eslint-disable-next-line @typescript-eslint/no-unused-vars
-            jest.spyOn(userService, "login").mockRejectedValue(ERROR);
+            jest.spyOn(userAuthService, "login").mockRejectedValue(ERROR);
 
             passportMock.mockImplementation(name => {
                 if (name !== "login") return;
